Reorganization & linting
