+++ src/main/java/com/github/springerris/db/KafRecord.java
@@ -1,4 +1,4 @@
-package com.github.springerris;
+package com.github.springerris.db;
 
 import org.jetbrains.annotations.NotNull;
